libpng errors almost solved
There was a warning with  incorrect sRGB profiles, is not a gamebreaking
bug, but is annoying. Some changes in libpng 1.6+ cause it issue warning
or even not work correctly with the original HP/MS sRGB profile. The old
profile uses a D50 whitepoint, where D65 is standard. This profile is
not uncommon, being used by Adobe Photoshop, although it was not
embedded into images by default.