You can click and kill a zergling now
Simply select a marine and right click onto the zergling you wish to
kill. The zergling dies at any given distance (we need to calculate the
minimum distance or range between marine and zergling).

GameManager needs to handle this kill.

Methods for deleting entities are done.

Methods implemented:

-KillEntity (overloaded both with a selection and an Entity)

-AddEntityToWave

SideNote: Next commits need to be done shorter and with a less extension
of time.