Refactored the code style of gameManager
Studied how ui works kinda and added a button too (not working at the
moment) in the game manager. I'm aware this should go into UI but nobody
is awake now to do it.

Tomorrow I'm gonna work along with R. Olasz to integrate winning and
losing labels and images, score visual feedback and stuff that happens
during the day, put score and wrapping up everything for release.

It's been a great day :)